\name{SIMEXBoost-package}
\alias{SIMEXBoost-package}
\docType{package}
\title{
Boosting Method for High-Dimensional Error-Prone Data
}
\description{
Implementation of the boosting procedure with the simulation and extrapolation (SIMEX) approach to address variable selection and estimation for high-dimensional data subject to measurement error in covariates.
}

\details{
This package aims to do variable selection and estimation by using the boosting procedure. An important and ubiquitous feature in the dataset is measurement error in covariates. To handle measurement error effects, we employ the simulation and extrapolation (SIMEX) method. In addition to commonly used generalized linear models (GLM), we also consider the accelerated failure time (AFT) model to fit length-biased and interval-censored survival data.}

\author{
Bangxu Qiu and Li-Pang Chen

Maintainer: Bangxu Qiu <1135427976@qq.com>
}



\seealso{
  \code{\link[SIMEXBoost]{SIMEXBoost}}
  %\code{\link[Metrics]{Metrics}}
}
\keyword{ package }

